Always use environment variables from add time + minor changes (#140)
* store environment variables at add time; made a bunch of things faster; better logging

* do not call scancel if the job list is empty

* optimize hash query, use multiprocessing for uploading src files

* remove redundant typing

* proper config updates in src reload

* use multithreading instead of multiprocessing

* cache reading file configs

* make sure that args are hashable

* fix setting experiments

* do not cache files from the venv

+++ src/seml/utils/slurm.py
@@ -174,6 +174,8 @@ def cancel_slurm_jobs(*job_ids: str | int, state: str | None = None):
     job_ids : Sequence[str]
         The job IDs of the jobs to cancel
     """
+    if len(job_ids) == 0:
+        return
     job_str = ' '.join(map(str, job_ids))
     if state is not None:
         subprocess.run(f'scancel -t {state} {job_str}', shell=True, check=False)
